+++ .eslintignore.errorfiles
@@ -1,11 +1,11 @@
 # autogenerated file: run scripts/generate-eslint-error-ignore-file to update.
 
 src/component-index.js
+src/components/structures/auth/ForgotPassword.js
 src/components/structures/BottomLeftMenu.js
 src/components/structures/CompatibilityPage.js
 src/components/structures/CreateRoom.js
 src/components/structures/LoggedInView.js
-src/components/structures/login/ForgotPassword.js
 src/components/structures/MessagePanel.js
 src/components/structures/NotificationPanel.js
 src/components/structures/RoomDirectory.js
@@ -15,6 +15,11 @@ src/components/structures/ScrollPanel.js
 src/components/structures/SearchBox.js
 src/components/structures/TimelinePanel.js
 src/components/structures/UploadBar.js
+src/components/views/auth/CountryDropdown.js
+src/components/views/auth/InteractiveAuthEntryComponents.js
+src/components/views/auth/PasswordLogin.js
+src/components/views/auth/RegistrationForm.js
+src/components/views/auth/ServerConfig.js
 src/components/views/avatars/BaseAvatar.js
 src/components/views/avatars/MemberAvatar.js
 src/components/views/create_room/RoomAlias.js
@@ -31,11 +36,6 @@ src/components/views/elements/UserSelector.js
 src/components/views/globals/MatrixToolbar.js
 src/components/views/globals/NewVersionBar.js
 src/components/views/globals/UpdateCheckBar.js
-src/components/views/login/CountryDropdown.js
-src/components/views/login/InteractiveAuthEntryComponents.js
-src/components/views/login/PasswordLogin.js
-src/components/views/login/RegistrationForm.js
-src/components/views/login/ServerConfig.js
 src/components/views/messages/MFileBody.js
 src/components/views/messages/RoomAvatarEvent.js
 src/components/views/messages/TextualBody.js
@@ -98,12 +98,12 @@ src/VectorConferenceHandler.js
 src/Velociraptor.js
 src/WhoIsTyping.js
 src/wrappers/withMatrixClient.js
-test/components/structures/login/Registration-test.js
+test/components/structures/auth/Registration-test.js
 test/components/structures/MessagePanel-test.js
 test/components/structures/ScrollPanel-test.js
 test/components/structures/TimelinePanel-test.js
+test/components/views/auth/RegistrationForm-test.js
 test/components/views/dialogs/InteractiveAuthDialog-test.js
-test/components/views/login/RegistrationForm-test.js
 test/components/views/rooms/MessageComposerInput-test.js
 test/components/views/rooms/RoomSettings-test.js
 test/mock-clock.js
